move out cli stuff, and base that on muxrpc-usage@3
